pyproject.toml
[tool.poetry]
name = "marketingnyt"
version = "0.1.0"
description = "SEO-optimized content site for marketingnyt.dk"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
wagtail = "^6.0"
psycopg2-binary = "^2.9"
redis = "^5.0"
django-redis = "^5.4"
whitenoise = "^6.6"
gunicorn = "^21.2"
pillow = "^10.2"
django-extensions = "^3.2"
python-dotenv = "^1.0"
beautifulsoup4 = "^4.12"
dj-database-url = "^2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^4.0"
black = "^24.0"
ruff = "^0.2"
factory-boy = "^3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"settings/*.py" = ["F405", "F403"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "marketingnyt.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing --reuse-db"
